+++ python/src/indian_fakedata/__init__.py
@@ -45,6 +45,7 @@ from indian_fakedata.utils.transliterate import (
 from indian_fakedata.utils.geo import generate_geo, state_geo_bounds, state_geo_anchor
 from indian_fakedata.utils.life_events import generate_life_events
 from indian_fakedata.utils.economy import generate_household_economy, emi_for
+from indian_fakedata.utils.festivals import generate_festivals
 from indian_fakedata.utils.schema import get_profile_schema, validate_profile
 from indian_fakedata.utils.privacy import strip_pii, PII_FIELDS
 
@@ -88,6 +89,7 @@ __all__ = [
     "generate_life_events",
     "generate_household_economy",
     "emi_for",
+    "generate_festivals",
     "get_profile_schema",
     "validate_profile",
     "strip_pii",